Only delete the mysql info if it exists

+++ app/Jobs/RemoveSite.php
@@ -33,8 +33,14 @@ class RemoveSite implements ShouldQueue
             return;
         }
 
-        $forge->deleteMysqlUser($project->forge_server_id, $branch->forge_mysql_user_id);
-        $forge->deleteMysqlDatabase($project->forge_server_id, $branch->forge_mysql_database_id);
+        if ($branch->forge_mysql_user_id !== null) {
+            $forge->deleteMysqlUser($project->forge_server_id, $branch->forge_mysql_user_id);
+        }
+
+        if ($branch->forge_mysql_database_id !== null) {
+            $forge->deleteMysqlDatabase($project->forge_server_id, $branch->forge_mysql_database_id);
+        }
+
         $forge->deleteSite($project->forge_server_id, $branch->forge_site_id);
     }
